% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Docker.R
\name{Docker}
\alias{Docker}
\title{Docker Commands Repository}
\description{
Docker Commands Repository

Docker Commands Repository
}
\concept{Docker Classes}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-show_running_containers}{\code{Docker$show_running_containers()}}
\item \href{#method-show_images}{\code{Docker$show_images()}}
\item \href{#method-remove_dangling_images}{\code{Docker$remove_dangling_images()}}
\item \href{#method-kill_all_containers}{\code{Docker$kill_all_containers()}}
\item \href{#method-kill_container}{\code{Docker$kill_container()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_running_containers"></a>}}
\if{latex}{\out{\hypertarget{method-show_running_containers}{}}}
\subsection{Method \code{show_running_containers()}}{
Show running containers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Docker$show_running_containers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_images"></a>}}
\if{latex}{\out{\hypertarget{method-show_images}{}}}
\subsection{Method \code{show_images()}}{
Show top level images, their repository, tags, and their size
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Docker$show_images()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_dangling_images"></a>}}
\if{latex}{\out{\hypertarget{method-remove_dangling_images}{}}}
\subsection{Method \code{remove_dangling_images()}}{
Remove layers that have no relationship to any tagged images
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Docker$remove_dangling_images()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kill_all_containers"></a>}}
\if{latex}{\out{\hypertarget{method-kill_all_containers}{}}}
\subsection{Method \code{kill_all_containers()}}{
Kill all running containers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Docker$kill_all_containers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kill_container"></a>}}
\if{latex}{\out{\hypertarget{method-kill_container}{}}}
\subsection{Method \code{kill_container()}}{
Kill a specific container
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Docker$kill_container(container_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{container_name}}{(\code{character}) container name}
}
\if{html}{\out{</div>}}
}
}
}
