% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DockerCompose.R
\name{DockerCompose}
\alias{DockerCompose}
\title{Use a docker-compose.yml File}
\description{
GIVEN a \code{docker-compose.yml},
WEHN \code{DockerCompose} is instantiated,
THEN the resulting object gives access to Docker commands.
}
\examples{

## ------------------------------------------------
## Method `DockerCompose$get`
## ------------------------------------------------

\donttest{\dontrun{DockerCompose$new()$get("shinyserver", "ports")}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DockerCompose$new()}}
\item \href{#method-get}{\code{DockerCompose$get()}}
\item \href{#method-start}{\code{DockerCompose$start()}}
\item \href{#method-stop}{\code{DockerCompose$stop()}}
\item \href{#method-restart}{\code{DockerCompose$restart()}}
\item \href{#method-reset}{\code{DockerCompose$reset()}}
\item \href{#method-browse_url}{\code{DockerCompose$browse_url()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a DockerCompose object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCompose$new(path_docker_compose = "./docker-compose.yml")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_docker_compose}}{(\code{character}) Path to docker-compose file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Get a value from a service
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCompose$get(service, field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{service}}{(\code{character}) Service name in \code{docker-compose.yml}.}

\item{\code{field}}{(\code{character}) Field name in \code{docker-compose.yml}.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{\dontrun{DockerCompose$new()$get("shinyserver", "ports")}}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
Create and start containers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCompose$start(service = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{service}}{(\code{character}) Service name in \code{docker-compose.yml}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop containers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCompose$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restart"></a>}}
\if{latex}{\out{\hypertarget{method-restart}{}}}
\subsection{Method \code{restart()}}{
Restart containers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCompose$restart(service = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{service}}{(\code{character}) Service name in \code{docker-compose.yml}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Stop and remove containers, networks, images and volumes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCompose$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse_url"></a>}}
\if{latex}{\out{\hypertarget{method-browse_url}{}}}
\subsection{Method \code{browse_url()}}{
Load URL into an HTML Browser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCompose$browse_url(service, slug = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{service}}{(\code{character}) Service name in \code{docker-compose.yml}.}

\item{\code{slug}}{(\code{character}) URL slug (e.g. \code{shiny-app-name}).}
}
\if{html}{\out{</div>}}
}
}
}
